refactor: 수강페이지 관련 로직 훅으로 분리
